api/requirements.txt
flask==3.0.3
mysql-connector-python==9.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9